interaction wrapper module
